+++ BOCCHI.Automator/IServiceCollectionExtensions.cs
@@ -27,6 +27,7 @@ public static class IServiceCollectionExtensions
         services.AddSingleton<IPotsTreasureMode, PotsTreasureService>();
         services.AddSingleton<Func<IPotsTreasureMode>>(sp => () => sp.GetRequiredService<IPotsTreasureMode>());
         services.AddSingleton<IllegalModeTreasureFillerService>();
+        services.AddSingleton<IIllegalModeStartableActivityProbe, IllegalModeStartableActivityProbe>();
         services.AddSingleton<TriageLatchService>();
         services.AddSingleton<PotTreasureHintTracker>();
         services.AddSingleton<IOnStart>(sp => sp.GetRequiredService<PotTreasureHintTracker>());